Refactor document loading logic into util module.
