+++ src/components/ChatBar.tsx
@@ -26,13 +26,12 @@ import {
   Tooltip,
   Typography,
 } from "@mui/material";
-import { WebPDFLoader } from "langchain/document_loaders/web/pdf";
 import Markdown from "markdown-to-jsx";
 import { v4 as uuidv4 } from "uuid";
 
 import { getContentConfig } from "../contentConfig";
 import { useThemeContext } from "../contexts/ThemeContext";
-import { DynamicFileLoader } from "../document_loaders/dynamic_file";
+import { getBase64Str } from "../document_loaders/util";
 import {
   CHAT_CONTAINER_HEIGHT_MAX,
   CHAT_CONTAINER_HEIGHT_MIN,
@@ -142,23 +141,7 @@ const ChatBar: React.FC = () => {
         const extensions = [".pdf"];
         const extension = "." + fileUploaded.name.split(".").pop() || "";
         if (extensions.includes(extension)) {
-          const loader = new DynamicFileLoader(fileUploaded, {
-            ".pdf": (file) => new WebPDFLoader(file, { splitPages: false }),
-          });
-          const docs = await loader.load();
-
-          // construct new base64 encoded string
-          let pageContent = "";
-          for (const doc of docs) {
-            pageContent += doc.pageContent + "\n\n";
-          }
-          const utf8Bytes = new TextEncoder().encode(pageContent);
-          let binary = "";
-          utf8Bytes.forEach((byte) => {
-            binary += String.fromCharCode(byte);
-          });
-
-          base64 = `data:${fileUploaded.type};base64,${btoa(binary)}`;
+          base64 = await getBase64Str(fileUploaded);
         }
 
         const attachment: Attachment = {
